fix(analyzer): health相关统计从int类型调整为BigDecimal
fix(analyzer): 修复部分空指针问题

+++ pokemon-crawler/src/main/java/com/mimosa/deeppokemon/analyzer/entity/event/DamageEventStat.java
@@ -8,5 +8,7 @@ package com.mimosa.deeppokemon.analyzer.entity.event;
 
 import com.mimosa.deeppokemon.analyzer.entity.EventTarget;
 
-public record DamageEventStat(EventTarget eventTarget, EventTarget damageOf, String damageFrom, int healthDiff) {
+import java.math.BigDecimal;
+
+public record DamageEventStat(EventTarget eventTarget, EventTarget damageOf, String damageFrom, BigDecimal healthDiff) {
 }
